Refactor database handling to use sqlx with SQLite
- Updated `Cargo.toml` in `ripclip-core` and `ripclipd` to replace rusqlite with sqlx.
- Modified `Clip` struct in `clip.rs` to derive `FromRow` for sqlx compatibility and changed `id` type from i32 to i64.
- Refactored `ClipRepository` to use `SqlitePool` for asynchronous database operations.
- Updated methods in `ClipRepository` to be asynchronous and adjusted their implementations accordingly.
- Changed the main function in `main.rs` to initialize `ClipRepository` with a database path and await the initialization.
- Refactored `ClipboardWorker` to handle asynchronous database operations when saving clips and fetching recent clips.
- Updated `IPCWorker` to handle asynchronous operations for client requests.
- Ensured proper error handling and logging for database operations.

+++ crates/ripclip-core/src/db/models/clip.rs
@@ -1,6 +1,8 @@
-#[derive(Debug)]
+use sqlx::FromRow;
+
+#[derive(Debug, FromRow)]
 pub struct Clip {
-    pub id: i32,
+    pub id: i64,
     pub content: String,
     pub copied_at: String,
 }
